componentize discussion voting modal
